new file:   www.rc 	deleted:    www/home-plugins/05_notifications/index.php 	modified:   www/home-plugins/05_notifications/notifications.php 	new file:   www/lib/htmlheaders/module-compatibility.php 	deleted:    www/lib/opt_htmlheaders/index.php 	modified:   www/lib/opt_htmlheaders/jquery.php 	deleted:    www/lib/range_icons/index.php 	deleted:    www/lib/shell/index.php 	modified:   www/lib/shell/list-interfaces.rc 	new file:   www/lib/themes/bright.css 	new file:   www/lib/themes/dark.css 	deleted:    www/login-plugins/01_internet-info/index.php 	modified:   www/login-plugins/01_internet-info/internet-info.php 	modified:   www/login-plugins/01_internet-info/shell.sh 	modified:   www/login-plugins/05_check-dash/check-dash.php 	deleted:    www/login-plugins/05_check-dash/index.php 	modified:   www/login-plugins/05_check-dash/shell.sh 	modified:   www/login-plugins/15_check-js/check-js.php 	deleted:    www/login-plugins/15_check-js/index.php 	modified:   www/net-wicd/index.php 	modified:   www/net-wicd/shell.php 	new file:   www/net-wicd/theme_bright.php 	modified:   www/net-wicd/wicd-config-injection.sh 	new file:   www/net-wifi/disabled 	modified:   www/net-wifi/index.php 	new file:   www/net-wifi/script.php 	modified:   www/net-wifi/shell.php 	modified:   www/net-wifi/shell.sh 	modified:   www/power/index.php 	modified:   www/power/menu-addon.php 	modified:   www/setup.sh 	deleted:    www/storage/description.php 	deleted:    www/storage/menu-addon.php 	modified:   www/sys-clock/index.php 	modified:   www/sys-clock/shell.php 	modified:   www/sys-notifications/index.php 	modified:   www/sys-sensors/index.php 	modified:   www/sys-sensors/shell.php 	renamed:    www/storage/config.php -> www/sys-storage/config.php 	new file:   www/sys-storage/description.php 	renamed:    www/storage/index.php -> www/sys-storage/index.php 	renamed:    www/storage/shell.sh -> www/sys-storage/shell.sh 	modified:   www/sys-updates/index.php 	modified:   www/sys-updates/menu-addon.php 	modified:   www/sys-updates/shell.php 	modified:   www/sys-users/index.php 	modified:   www/sys-users/shell.sh

+++ www/setup.sh
@@ -2,52 +2,57 @@
 
 echo '05_notifications'
 cd home-plugins/05_notifications
-rm index.php
+[ -e index.php ] && rm index.php
 ln -s ../../lib/prevent-index.php index.php
 chmod 755 shell.sh
 
 echo 'range_icons'
 cd ../../lib/range_icons
-rm index.php
+[ -e index.php ] && rm index.php
 ln -s ../prevent-index.php index.php
 
 echo 'opt_htmlheaders'
 cd ../opt_htmlheaders
-rm index.php
+[ -e index.php ] && rm index.php
 ln -s ../prevent-index.php index.php
 
 echo 'shell libs'
 cd ../shell
-rm index.php
+[ -e index.php ] && rm index.php
 ln -s ../prevent-index.php index.php
 
 echo '01_internet-info'
 cd ../../login-plugins/01_internet-info
-rm index.php
+[ -e index.php ] && rm index.php
 ln -s ../../lib/prevent-index.php index.php
 chmod 755 shell.sh
 
 echo '05_check-dash'
 cd ../05_check-dash
-rm index.php
+[ -e index.php ] && rm index.php
 ln -s ../../lib/prevent-index.php index.php
 chmod 755 shell.sh
 
 echo '15_check-js'
 cd ../15_check-js
-rm index.php
+[ -e index.php ] && rm index.php
 ln -s ../../lib/prevent-index.php index.php
 
 echo 'net-wifi'
 cd ../../net-wifi
 chmod 755 shell.sh
 
+echo 'net-wicd'
+cd ../net-wicd
+chmod 755 shell.sh
+chmod 755 wicd-config-injection.sh
+
 echo 'power'
 cd ../power
 chmod 755 shell.sh
 
-echo 'storage'
-cd ../storage
+echo 'sys-storage'
+cd ../sys-storage
 chmod 755 shell.sh
 
 echo 'sys-notifications'
@@ -66,8 +71,72 @@ echo 'sys-users'
 cd ../sys-users
 chmod 755 shell.sh
 
-echo 'setup.sh'
 cd ..
+while true; do
+	error=false
+	echo; echo -n 'download jquery? [y/n] '
+	read answer
+	if [ "$answer" = 'y' ]; then
+		cd lib
+		echo 'jquery-3.3.1.min.js'
+		wget https://code.jquery.com/jquery-3.3.1.min.js > /dev/null 2>&1 && ln -s jquery-3.3.1.min.js jquery.js || error=true
+		echo 'jquery-1.9.1.min.js'
+		wget http://code.jquery.com/jquery-1.9.1.min.js > /dev/null 2>&1 && ln -s jquery-1.9.1.min.js jquery-old.js || error=true
+		if $error; then
+			while true; do
+				echo; echo -n 'download failed, create download script? [y/n] '
+				read answerb
+				if [ "$answerb" = 'y' ]; then
+					cd ..
+					echo '#!/bin/sh' > download-jquery.sh
+					echo 'error=false' >> download-jquery.sh
+					echo 'cd lib' >> download-jquery.sh
+					echo 'echo "jquery-3.3.1.min.js"' >> download-jquery.sh
+					echo 'wget https://code.jquery.com/jquery-3.3.1.min.js > /dev/null 2>&1 && ln -s jquery-3.3.1.min.js jquery.js || error=true' >> download-jquery.sh
+					echo 'echo "jquery-1.9.1.min.js"' >> download-jquery.sh
+					echo 'wget http://code.jquery.com/jquery-1.9.1.min.js > /dev/null 2>&1 && ln -s jquery-1.9.1.min.js jquery-old.js || error=true' >> download-jquery.sh
+					echo 'cd ..' >> download-jquery.sh
+					echo '$error && echo "download error!"' >> download-jquery.sh
+					echo '$error || rm download-jquery.sh' >> download-jquery.sh
+					echo 'exit 0' >> download-jquery.sh
+					chmod 755 download-jquery.sh
+					cd lib
+					break
+				fi
+				[ "$answerb" = 'n' ] && break
+			done
+		fi
+		cd ..
+		break
+	fi
+	[ "$answer" = 'n' ] && break
+done
+
+while true; do
+	echo; echo -n 'remove deprecated net-wifi module? [y/n] '
+	read answer
+	if [ "$answer" = 'y' ]; then
+		echo 'rm net-wifi'
+		rm -r net-wifi
+		break
+	fi
+	[ "$answer" = 'n' ] && break
+done
+
+while true; do
+	echo; echo -n 'remove module-compatibility header? [y/n] '
+	read answer
+	if [ "$answer" = 'y' ]; then
+		echo 'rm module-compatibility.php'
+		cd lib
+		rm -r htmlheaders
+		cd ..
+		break
+	fi
+	[ "$answer" = 'n' ] && break
+done
+
+echo; echo 'setup.sh'
 rm setup.sh
 
 echo; echo 'OK'
